==> 2-primary-analyses.Rmd <==
---
title: "TBS for treating depression in elderly"
author: "Matthias Lüthi"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    number_sections: true
    toc_float: true
editor_options: 
  markdown: 
    wrap: 72
---

```{r setup, include=FALSE}
# Only show figures and tables:
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
```

```{r load-libraries-data}
library(tidyverse)
library(lme4)
library(lmerTest)
library(lmtest)
library(broom.mixed)
library(effectsize)
library(interactions)
library(flextable)
library(ftExtra)
library(ggpubr)
library(MASS)
library(here)
library(nlme)
library(writexl)
library(infer)
library(BI)

select <- dplyr::select

# Load custom functions
source("0-custom-functions.R")

load(here("Data", "elderly_cleaned_wide.RData"))
load(here("Data", "elderly_cleaned_long.RData"))
```

```{r create-additional-dfs}
dss_bl <- filter(dss, time == 0)

summaries <- dss %>%
  group_by(time, condition_rev) %>%  
  summarise(across(c(hdrs_total, madrs_total, 
                     gds_total, ymrs_total, 
                     panas_pos_total, panas_neg_total, CGI1), 
                   list(Mean = ~mean(.x, na.rm = TRUE), 
                        SD = ~sd(.x, na.rm = TRUE), 
                        SEM = ~sd(.x, na.rm = TRUE)/sqrt(sum(!is.na(.x))),
                        N = ~sum(!is.na(.x))
                   ),
                   .names = "{col}_{fn}"),
            across(hdrs_response:madrs_remission,
                   ~sum(.x, na.rm=TRUE)),
            # hdrs_response_prcnt = 
            across(hdrs_response:madrs_remission,
                   ~sum(.x, na.rm=TRUE)/sum(!is.na(.x)),
                   .names = "{col}_Prcnt")
  ) %>%
  ungroup()

dss_wide <- dss_wide %>%
  mutate(guess_patient = as.factor(case_when(
    Ceg_Paciente1_12 == 0 ~ "Guess sham",
    Ceg_Paciente1_12 == 1 ~ "Guess active",
    Ceg_Paciente1_12 == 2 ~ NA)),
    guess_rater = as.factor(case_when(
      Ceg_Avaliador2_12 == 0 ~ "Guess sham",
      Ceg_Avaliador2_12 == 1 ~ "Guess active")),
    adverse_effects_w12 = Ceg_Paciente3_12) %>%
  rowwise() %>%
  mutate(adverse_effects_total = mean(c(Ceg_Paciente3_1, Ceg_Paciente3_2,
           Ceg_Paciente3_4, Ceg_Paciente3_6, Ceg_Paciente3_8, Ceg_Paciente3_12), 
           na.rm=TRUE)) %>%
  ungroup()
```

# Tables 
## Table 1
```{r table-1, message = FALSE}

# Prepare data frame
table1 <- data.frame(matrix(nrow = 0, ncol = 4))
conditions_with_ns <- mapply(function(x, y) paste0(x, "---N = ", y), 
                             levels(dss_bl$condition_rev), 
                             dss_bl %>%
                               count(condition_rev) %>%
                               pull(n))
colnames(table1) <- c("Characteristic",
                      conditions_with_ns, 
                      "P")

sample_ns <- dss_bl %>% count(condition_rev) %>% pull(n) %>% unique()

vars <- list(c("Demographics", "", "header row"),
             c("gender", "Gender (% Female)", "cat", "female"),
             c("Idade", "Age, y", "cont"),
             c("etnia", "White - no. (%)", "cat", 1),
             c("anos_estudo", "Years of education - mean (SD)", "cont"),
             c("renda", "At least 3 minimum wages - no. (%)", "cat", 2, 3, 4),
             c("ocupacao", "Retired - no. (%)", "cat", 4),
             c("relationship", "Relationship Status (% Partnered)", "cat", "yes"),
             c("physical_activity", "Physical Activity (% Active)", "cat", "high"),
             c("Clinical characteristics", "", "header row"),
             c("hipertensao", "Hypertension - no. (%)", "cat", 1),
             c("hipotireoidismo", "Hypothyroidism - no. (%)", "cat", 1),
             c("mini_28", "Duration of current episode, months - mean (SD)", "cont"),
             c("mini_69", "Refractory depression - no. (%)", "cat", "1"),
             c("in_psychotherapy", "Currently in psychotherapy - no. (%)", "cat", TRUE),
             c("Benzodiazepines", "Current benzodiazepine use - no. (%)", "cat", "1"),
             c("mini_73_total", "No. failed AD trials in lifetime - median (IQR)", "count"),
             c("mini_26", " Onset age of MDD, years - mean (SD)", "cont"),
             c("mini_27", "No. previous depressive episodes - median (IQR)", "count"),
             c("Scales at baseline", "", "header row"),
             c("hdrs_total", "Hamilton Depression Rating Scale (HDRS-17)", "cont"),
             c("madrs_total", "Montgomery-Asberg Depression Rating Scale (MADRS)", "cont"),
             c("gds_total", "Geriatric Depression Scale (GDS)", "cont"),
             c("ymrs_total", "Mania (YMRS)", "cont"),
             c("panas_pos_total", "Positive affect (PANAS)", "cont"),
             c("panas_neg_total", "Negative affect (PANAS)", "cont"),
             c("CGI1", "Clinical Global Impression (CGI)", "cont"),
             c("G_CIRS_0", "Cumulative Illness (CIRS-G)", "cont")
)

for (var in vars) {
  var_name <- var[1]
  var_desc <- var[2]
  var_type <- var[3]

  if (grepl(var_type, "header row")) {
    table1[nrow(table1)+1, ] <- c(var_name, "", "", NA_real_)
    
  } else if (grepl(var_type, "cont")) {

    ns <- dss_bl %>%
      group_by(condition_rev) %>%
      summarize(desc = paste0(round(mean(.data[[var_name]], na.rm=TRUE), 1),
                              " ± ",
                              round(sd(.data[[var_name]], na.rm=TRUE), 1),
                              ifelse(
                                sum(!is.na(.data[[var_name]])) %in% sample_ns,
                                "",
                                paste0("; n=", sum(!is.na(.data[[var_name]]))))
      )) %>%
      pull(desc)

    formula <- as.formula(paste(var_name, "~ condition"))

    table1[nrow(table1)+1, ] <- c(
      var_desc,
      ns,
      tidy(aov(formula, dss_bl))$p.value[1]
    )
    
  } else if (grepl(var_type, "cat")) {
    
    relevant_cats <- var[4:length(var)] 
    el_temp <- dss_bl %>%
      mutate(current_var = if_else(is.na(.data[[var_name]]), NA,
                                  .data[[var_name]] %in% relevant_cats)
             )%>%
      select(condition_rev, current_var)
  
    ns <- el_temp %>%
      group_by(condition_rev) %>%
      summarize(desc = paste0(sum(current_var, na.rm=TRUE),
                              " (",
                              round(sum(current_var, na.rm=TRUE)/sum(!is.na(current_var))*100),
                              ")",
                              ifelse(
                                sum(!is.na(current_var)) %in% sample_ns,
                                "",
                                paste0("; n=", sum(!is.na(current_var))))
      )) %>%
      pull(desc)

    table1[nrow(table1)+1, ] <- c(
      var_desc,
      ns,
      test_independence(el_temp, "condition_rev", "current_var")
    )

  } else if (grepl(var_type, "count")) {
  
    ns <- dss_bl %>%
      group_by(condition_rev) %>%
      summarize(desc = paste0(median(.data[[var_name]], na.rm=TRUE),
                              " (",
                              round(quantile(.data[[var_name]], na.rm=TRUE, probs = 0.25)),
                              "-",
                              round(quantile(.data[[var_name]], na.rm=TRUE, probs = 0.75)),
                              ")",
                              ifelse(
                                sum(!is.na(.data[[var_name]])) %in% sample_ns,
                                "",
                                paste0("; n=", sum(!is.na(.data[[var_name]]))))
      )
      ) %>%
      pull(desc)

    table1[nrow(table1)+1, ] <- c(
      var_desc,
      ns,
      test_count(dss_bl, var_name, "condition_rev")
    )
  }
}


ft_table1  <- table1 %>%
  mutate(P = ifelse(P == "-", 100, round(as.numeric(P), 3)),
         P = case_when(
           P == 100 ~ "-",
           P == 1 ~ ">0.99",
           P == 0 ~ "<0.001",
           P > 0.02 ~ as.character(round(P, 2)),
           TRUE ~ as.character(P))
  ) %>%
  flextable() %>%
  separate_header(sep = "---") %>%
  italic(i = is.na(table1$P), j = 1) %>%
  autofit()
ft_table1

ft_table1_pub  <- table1 %>%
  select(-P) %>%
  mutate(across(
    `Active TBS---N = 54`:`Sham TBS---N = 54`,
    ~ str_replace(., ";.*", "*"))
  ) %>%
  flextable() %>%
  italic(i = is.na(table1$P), j = 1) %>%
  separate_header(sep = "---") %>%
  fit_pagewidth()


save_as_docx(ft_table1, path = here("Output", "Table1_Idosos.docx"))
save_as_docx(ft_table1_pub, path = here("Output", "Table1_Idosos_Pub.docx"))

```
*Note:* (*) indicates that data from the HERAH system is missing (33 of 108 
patients). There were no cases of hyperthyroidism. 

## Scale scores across time 
```{r}
# This does not include CIRS-G, as it was only collected at baseline!

table_scales <- summaries %>%
  pivot_longer(cols = contains("_total_"),
               names_pattern = "^(.*)_total_(.*)$",
               names_to = c("Scale", ".value")) %>%
  arrange(desc(Scale), time) %>%
  pivot_wider(id_cols = c("condition_rev", "Scale"), 
              names_from = "time",
              values_from = c("Mean", "SD"),
              names_sep = "_"         
  ) %>%
  mutate(`Base-line` = paste(round(Mean_0, 2), "±", round(SD_0, 2)),
         `Week 1` = paste(round(Mean_1, 2), "±", round(SD_1, 2)),
         `Week 2` = paste(round(Mean_2, 2), "±", round(SD_2, 2)),
         `Week 4` = paste(round(Mean_4, 2), "±", round(SD_4, 2)),
         `Week 6` = paste(round(Mean_6, 2), "±", round(SD_6, 2)),
         `Week 8` = paste(round(Mean_8, 2), "±", round(SD_8, 2)),
         `Week 12` = paste(round(Mean_12, 2), "±", round(SD_12, 2)),
         
         Scale = case_when(
           Scale == "panas_pos" ~ "Positive Affect PANAS",
           Scale == "panas_neg" ~ "Negative Affect PANAS",
           Scale == "hdrs" ~ "HDRS-17",
           TRUE ~ toupper(Scale)
         )) %>%
  select(Scale, Group = condition_rev, `Base-line`:`Week 12`)

# Define scale order
custom_order <- c("HDRS-17", "MADRS", "GDS", "YMRS",
                  "Positive Affect PANAS", "Negative Affect PANAS") 

# Reorder the dataframe based on custom order
table_scales <- table_scales %>%
  mutate(Scale = factor(Scale, levels = custom_order)) %>%
  arrange(Scale)

table_cgi <- summaries %>%
  pivot_longer(cols = contains("CGI1") ,
               names_pattern = "^(CGI1)_(.*)$",
               names_to = c("Scale", ".value")) %>%
  arrange(desc(Scale), time) %>%
  pivot_wider(id_cols = c("condition_rev", "Scale"), 
              names_from = "time",
              values_from = c("Mean", "SD"),
              names_sep = "_"         
  ) %>%
  mutate(`Base-line` = paste(round(Mean_0, 2), "±", round(SD_0, 2)),
         `Week 1` = paste(round(Mean_1, 2), "±", round(SD_1, 2)),
         `Week 2` = paste(round(Mean_2, 2), "±", round(SD_2, 2)),
         `Week 4` = paste(round(Mean_4, 2), "±", round(SD_4, 2)),
         `Week 6` = paste(round(Mean_6, 2), "±", round(SD_6, 2)),
         `Week 8` = paste(round(Mean_8, 2), "±", round(SD_8, 2)),
         `Week 12` = paste(round(Mean_12, 2), "±", round(SD_12, 2)),
         
         Scale = "CGI") %>%
  select(Scale, Group = condition_rev, `Base-line`:`Week 12`)

table_scales <- bind_rows(table_scales, table_cgi)
  
table_scales %>%
  filter(Scale == "HDRS-17") %>%
  select(-Scale) %>%
  flextable() %>%
  fit_pagewidth() %>%
  save_as_docx(path = here("Output", "Table_HDRS17_timepoints.docx"))

table_scales %>%
  flextable() %>%
  fit_pagewidth() %>%
  save_as_docx(path = here("Output", "Table_Scales_timepoints.docx"))
```

## Response & Remission 
```{r}
resp_rem <- dss %>%
  group_by(condition_rev, time) %>%
  summarize(
    "Response - no. (%)" = paste0(
      sum(hdrs_response, na.rm=T), 
      " (", 
      round(sum(hdrs_response, na.rm=T)/sum(!is.na(hdrs_response))*100),
      ")"
    ),
    "Remission - no. (%)" = paste0(
      sum(hdrs_remission, na.rm=T), 
      " (", 
      round(sum(hdrs_remission, na.rm=T)/sum(!is.na(hdrs_remission))*100),
      ")"
    )
  ) %>%
  pivot_wider(id_cols = condition_rev,
              names_from = time,
              values_from = c("Response - no. (%)",
                              "Remission - no. (%)"),
              names_sep = "_"
  ) %>%
  pivot_longer(cols = "Response - no. (%)_0":"Remission - no. (%)_12",
               names_sep = "_",
               names_to = c("Type", ".value")
  ) %>%
  arrange(condition_rev, desc(Type)) %>%
  select(-`0`)

colnames(resp_rem) <- c("Group", "Type", "Week 1", "Week 2", "Week 4",
                        "Week 6", "Week 8", "Week 12")

ft_resp_rem <- resp_rem %>% 
  flextable() %>%
  align(j =  3:6, align = "right", part = "all") %>%
  autofit()

ft_resp_rem

save_as_docx(fit_pagewidth(ft_resp_rem), 
             path = here("Output", "Table_ResponseRemission_timepoints.docx")
             )
```

## Hamilton Depression Rating Scale (HDRS) score reductions
```{r}
table_diffs <- dss_wide %>%
  group_by(condition_rev) %>%
  summarize(mean(diff_hdrs_6, na.rm=TRUE),
            sd(diff_hdrs_6, na.rm=TRUE),
            mean(diff_hdrs_12, na.rm=TRUE),
            sd(diff_hdrs_12, na.rm=TRUE))

table_diffs

```


# Figures 
All error bars represent 95% confidence intervals. 

## Primary and secondary outcomes
```{r longitudinal-figures-recalculated}
ggplot(summaries, aes(x = time, y = hdrs_total_Mean, 
                       color = condition_rev, shape = condition_rev)) +
  geom_line(size = 1.2) +
  geom_point(size = 3.3) +
  geom_errorbar(aes(ymin = hdrs_total_Mean - 1.96*hdrs_total_SEM,
                    ymax = hdrs_total_Mean + 1.96*hdrs_total_SEM),
                width = 0.7,
                size = 0.75,
                alpha = 0.8
  ) +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
                     labels = sapply(c("Base- line", 1, 2, 4, 6, 8, 12),
                                     function(x) str_wrap(x, width = 5))) +
  expand_limits(y = 0) +
  labs(y = "HDRS-17", x = "Week") +
  theme_nejm()


ggplot(summaries, aes(x = time, y = madrs_total_Mean, 
                      color = condition_rev, shape = condition_rev)) +
  geom_line(size = 1.2) +
  geom_point(size = 3.3) +
  geom_errorbar(aes(ymin = madrs_total_Mean - 1.96*madrs_total_SEM,
                    ymax = madrs_total_Mean + 1.96*madrs_total_SEM),
                width = 0.7,
                size = 0.75,
                alpha = 0.8
  ) +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
                     labels = sapply(c("Base- line", 1, 2, 4, 6, 8, 12),
                                     function(x) str_wrap(x, width = 5))) +
  expand_limits(y = 0) +
  labs(y = "MADRS", x = "Weeks") +
  theme_nejm()

ggplot(filter(summaries, time != 8), 
       aes(x = time, y = gds_total_Mean, 
           color = condition_rev, shape = condition_rev)) +
  geom_line(size = 1.2) +
  geom_point(size = 3.3) +
  geom_errorbar(aes(ymin = gds_total_Mean - 1.96*gds_total_SEM,
                    ymax = gds_total_Mean + 1.96*gds_total_SEM),
                width = 0.7,
                size = 0.75,
                alpha = 0.8
  ) +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 12),
                     labels = sapply(c("Base- line", 1, 2, 4, 6, 12),
                                     function(x) str_wrap(x, width = 5))) +
  expand_limits(y = 0) +
  labs(y = "GDS", x = "Weeks") +
  theme_nejm()

ggplot(summaries, aes(x = time, y = panas_neg_total_Mean, 
                      color = condition_rev, shape = condition_rev)) +
  geom_line(size = 1.2) +
  geom_point(size = 3.3) +
  geom_errorbar(aes(ymin = panas_neg_total_Mean - 1.96*panas_neg_total_SEM,
                    ymax = panas_neg_total_Mean + 1.96*panas_neg_total_SEM),
                width = 0.7,
                size = 0.75,
                alpha = 0.8
  ) +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
                     labels = sapply(c("Base- line", 1, 2, 4, 6, 8, 12),
                                     function(x) str_wrap(x, width = 5))) +
  expand_limits(y = 0) +
  labs(y = "Negative affect (PANAS)", x = "Weeks") +
  theme_nejm()

ggplot(summaries, aes(x = time, y = panas_pos_total_Mean, 
                      color = condition_rev, shape = condition_rev)) +
  geom_line(size = 1.2) +
  geom_point(size = 3.3) +
  geom_errorbar(aes(ymin = panas_pos_total_Mean - 1.96*panas_pos_total_SEM,
                    ymax = panas_pos_total_Mean + 1.96*panas_pos_total_SEM),
                width = 0.7,
                size = 0.75,
                alpha = 0.8
  ) +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
                     labels = sapply(c("Base- line", 1, 2, 4, 6, 8, 12),
                                     function(x) str_wrap(x, width = 5))) +
  expand_limits(y = 0) +
  labs(y = "Positive affect (PANAS)", x = "Weeks") +
  theme_nejm()

ggplot(summaries, aes(x = time, y = ymrs_total_Mean, 
                      color = condition_rev, shape = condition_rev)) +
  geom_line(size = 1.2) +
  geom_point(size = 3.3) +
  geom_errorbar(aes(ymin = ymrs_total_Mean - 1.96*ymrs_total_SEM,
                    ymax = ymrs_total_Mean + 1.96*ymrs_total_SEM),
                width = 0.7,
                size = 0.75,
                alpha = 0.8
  ) +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
                     labels = sapply(c("Base- line", 1, 2, 4, 6, 8, 12),
                                     function(x) str_wrap(x, width = 5))) +
  expand_limits(y = 0) +
  labs(y = "YMRS", x = "Weeks") +
  theme_nejm()

ggplot(summaries, aes(x = time, y = CGI1_Mean, 
                      color = condition_rev, shape = condition_rev)) +
  geom_line(size = 1.2) +
  geom_point(size = 3.3) +
  geom_errorbar(aes(ymin = CGI1_Mean - 1.96*CGI1_SEM,
                    ymax = CGI1_Mean + 1.96*CGI1_SEM),
                width = 0.7,
                size = 0.75,
                alpha = 0.8
  ) +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
                     labels = sapply(c("Base- line", 1, 2, 4, 6, 8, 12),
                                     function(x) str_wrap(x, width = 5))) +
  expand_limits(y = 0) +
  labs(y = "CGI", x = "Weeks") +
  theme_nejm()
```

## Alternative figures for main outcome
```{r}
# absolute values
ggplot(summaries, aes(x = time, y = hdrs_total_Mean, 
                      color = condition_rev, shape = condition_rev)) +
  geom_line(size = 1.25) +
  geom_point(size = 3.5) +
  geom_errorbar(
    aes(ymin = hdrs_total_Mean - hdrs_total_SD,
        ymax = hdrs_total_Mean + hdrs_total_SD),
                width = 0.5,
                # alpha = 0.8,
                size = 1
                
  ) +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
                     labels = sapply(c("Base- line", 1, 2, 4, 6, 8, 12),
                                     function(x) str_wrap(x, width = 5))) +
  expand_limits(y = 0) +
  scale_color_manual(values = c("Sham TBS" = "blue", "Active TBS" = "red")) +
  labs(y = "HDRS-17", x = "Week") +
  theme_pub()

ggsave("Lineplot_HDRS_SD.pdf", width = 8, height = 6, units = "in")

# Percent change 
summaries_prcnt <- dss %>%
  group_by(subject) %>%
  mutate(
    hdrs_prcnt = ((hdrs_total - hdrs_total[time == 0]) / hdrs_total[time == 0]) * 100
    ) %>%
  ungroup() %>%
  group_by(time, condition_rev) %>%
  summarise(
    hdrs_prcnt_mean = mean(hdrs_prcnt, na.rm = TRUE), 
    hdrs_prcnt_sd = sd(hdrs_prcnt, na.rm = TRUE), 
    hdrs_prcnt_sem = sd(hdrs_prcnt, na.rm = TRUE)/sqrt(sum(!is.na(hdrs_prcnt)))
  )
    
ggplot(summaries_prcnt, aes(x = time, y = hdrs_prcnt_mean, 
                      color = condition_rev, shape = condition_rev)) +
  geom_line(size = 1.25) +
  geom_point(size = 3.5) +
  geom_errorbar(
        aes(ymin = hdrs_prcnt_mean - hdrs_prcnt_sem,
        ymax = hdrs_prcnt_mean + hdrs_prcnt_sem),
                width = 0.6,
                # alpha = 0.8,
                size = 1
                
  ) +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
                     labels = sapply(c("Base- line", 1, 2, 4, 6, 8, 12),
                                     function(x) str_wrap(x, width = 5))) +
  expand_limits(y = -100) +
  scale_y_continuous(labels = function(x) paste0(x, "%")) +
  labs(y = "Percent Change") +
  labs(y = "HDRS-17 change", x = "Week") +
  theme_pub()

```


## Response and Remission
```{r response-remission}
resp_rem <- dss %>%
  group_by(condition, time) %>%
  summarize(
    "HDRS response - no. (%)" = paste0(
      sum(hdrs_response, na.rm=T), 
      " (", 
      round(sum(hdrs_response, na.rm=T)/sum(!is.na(hdrs_response))*100),
      ")"
      ),
    "HDRS remission - no. (%)" = paste0(
      sum(hdrs_remission, na.rm=T), 
      " (", 
      round(sum(hdrs_remission, na.rm=T)/sum(!is.na(hdrs_remission))*100),
      ")"
      ),
    "MADRS response - no. (%)" = paste0(
      sum(madrs_response, na.rm=T), 
      " (", 
      round(sum(madrs_response, na.rm=T)/sum(!is.na(madrs_response))*100),
      ")"
      ),
    "MADRS remission - no. (%)" = paste0(
      sum(madrs_remission, na.rm=T), 
      " (", 
      round(sum(madrs_remission, na.rm=T)/sum(!is.na(madrs_remission))*100),
      ")"
      )
  ) %>%
  pivot_wider(id_cols = condition,
               names_from = time,
               values_from = c("HDRS response - no. (%)",
                               "HDRS remission - no. (%)",
                               "MADRS response - no. (%)",
                               "MADRS remission - no. (%)"),
               names_sep = "_"
  ) %>%
  pivot_longer(cols = "HDRS response - no. (%)_0":"MADRS remission - no. (%)_12",
               names_sep = "_",
               names_to = c("Type", ".value")
  ) %>%
  arrange(Type, desc(condition))

colnames(resp_rem) <- c("Condition", "Type", "Baseline", "Week 1", "Week 2",
                        "Week 4", "Week 6", "Week 8", "Week 12")
ft_resp_rem <- resp_rem %>% 
  flextable() %>%
  autofit()

ft_resp_rem

save_as_docx(ft_resp_rem, path = "TBSinElderly_ResponseRemission.docx")

# Figure for Week 12
dss %>%
  filter(time == 12) %>%
  group_by(condition_rev) %>%
  summarize(
         Response = sum(hdrs_response, na.rm = TRUE)/sum(!is.na(hdrs_response))*100,
         Remission = sum(hdrs_remission, na.rm = TRUE)/sum(!is.na(hdrs_remission))*100
         ) %>%
  pivot_longer(cols = c(Response, Remission),
               names_to = "Type",
               values_to = "Prcnt") %>%
  ggplot(aes(x = condition_rev, y = Prcnt, fill = condition_rev)) +
  geom_bar(position="stack", stat="identity", show.legend = FALSE) +
  geom_text(aes(label = sapply(round(Prcnt, 0), function(x) paste0(x, "%"))), 
            nudge_y = 4, size = 4) +
  labs(y = "Percent", x = "", title = "Evaluation at week 12") +
  scale_y_continuous(limits = c(0, 100)) +
  scale_x_discrete(labels = function(x) 
    stringr::str_wrap(x, width = 5))+
  facet_wrap(~Type) +
  theme_nejm()

# Figure for Week 6
dss %>%
  filter(time == 6) %>%
  group_by(condition_rev) %>%
  summarize(
         Response = sum(hdrs_response, na.rm = TRUE)/sum(!is.na(hdrs_response))*100,
         Remission = sum(hdrs_remission, na.rm = TRUE)/sum(!is.na(hdrs_remission))*100
         ) %>%
  pivot_longer(cols = c(Response, Remission),
               names_to = "Type",
               values_to = "Prcnt") %>%
  ggplot(aes(x = condition_rev, y = Prcnt, fill = condition_rev)) +
  geom_bar(position="stack", stat="identity", show.legend = FALSE) +
  geom_text(aes(label = sapply(round(Prcnt, 0), function(x) paste0(x, "%"))), 
            nudge_y = 4, size = 4) +
  labs(y = "Percent", x = "", title = "Evaluation at week 6") +
  scale_y_continuous(limits = c(0, 100)) +
  scale_x_discrete(labels = function(x) 
    stringr::str_wrap(x, width = 5))+
  facet_wrap(~Type) +
  theme_nejm()

## Same figure as above, but in Portuguese
# dss %>%
#   filter(time == 12) %>%
#   group_by(condition) %>%
#   summarize(
#          Resposta = sum(hdrs_response, na.rm = TRUE)/sum(!is.na(hdrs_response))*100,
#          Remissão = sum(hdrs_remission, na.rm = TRUE)/sum(!is.na(hdrs_remission))*100
#          ) %>%
#   pivot_longer(cols = c(Resposta, Remissão),
#                names_to = "Type",
#                values_to = "Prcnt") %>%
#   ggplot(aes(x = condition, y = Prcnt, fill = condition)) +
#   geom_bar(position="stack", stat="identity", show.legend = FALSE) +
#   geom_text(aes(label = sapply(round(Prcnt, 0), function(x) paste0(x, "%"))), 
#             nudge_y = 4, size = 4) +
#   scale_fill_brewer(palette = "Dark2", direction = 1) +
#   labs(y = "Percentagem", x = "", title = "Avaliação na semana 12") +
#   scale_y_continuous(limits = c(0, 100)) +
#   scale_x_discrete(labels = c("Placebo", "TBS", "Placebo", "TBS"))+
#   # scale_x_discrete(labels = function(x) 
#   #   stringr::str_wrap(x, width = 5))+
#   facet_wrap(~Type) +
#   theme_nejm()
```


```{r barplots-resp-rem}

ggplot(summaries, aes(x = time, y = hdrs_response, 
                      color = condition_rev, fill = condition_rev)) +
  geom_col(position = "dodge") +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
  ) +
  expand_limits(y = 0) +
  labs(y = "HDRS Response", x = "Weeks") +
  theme_nejm()

ggplot(summaries, aes(x = time, y = hdrs_remission, 
                      color = condition_rev, fill = condition_rev)) +
  geom_col(position = "dodge") +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
  ) +
  expand_limits(y = 0) +
  labs(y = "HDRS Remission", x = "Weeks") +
  theme_nejm()

ggplot(summaries, aes(x = time, y = madrs_response, 
                      color = condition_rev, fill = condition_rev)) +
  geom_col(position = "dodge") +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
  ) +
  expand_limits(y = 0) +
  labs(y = "MADRS Response", x = "Weeks") +
  theme_nejm()

ggplot(summaries, aes(x = time, y = madrs_remission, 
                      color = condition_rev, fill = condition_rev)) +
  geom_col(position = "dodge") +
  scale_x_continuous(breaks = c(0, 1, 2, 4, 6, 8, 12),
  ) +
  expand_limits(y = 0) +
  labs(y = "MADRS Remission", x = "Weeks") +
  theme_nejm()

```

MADRS response and remission rates reflect the temporary drop in scores at week 1.

# Antidepressive effects of treatment

## Linear mixed models

```{r lmm-hdrs}
model_hdrs <- lmer(hdrs_total ~ condition*time_log + (1|subject), 
                    data = dss)
summary(model_hdrs)

model_tidy <- tidy(model_hdrs)[1:4, ]
                   
# Calculate effect size
cohensd <- t_to_d(model_tidy$statistic, model_tidy$df) 
model_tidy <- cbind(model_tidy, cohensd)
model_tidy
```



A linear mixed-effects model was used to investigate if treatment was effective.
Time was assumed to be linear on a natural logarithmic scale. 
HDRS-17 was the outcome, group and time were included as interacting fixed 
effects, and subject as random-intercept effect. 
Results showed a reduction in HDRS-17 scores across time, independent of group
(t(606) = -10.89, P < .001). 
There was a a significant interaction of group and time 
(t(604) = -2.01, P = 0.045). This indicates that active TBS treatment was 
effective in reducing depressive symptoms. 

The following tables show the interaction tests for all outcomes at different 
endpoints (week 6 or 12). 

```{r function-definitions-for-lmer}
run_lmer <- function(df, outcome, week_max, print = FALSE) {
    formula <- as.formula(paste(outcome, "~ condition*time_log + (1|subject)"))

    lmm <- lmer(formula, filter(df, time <= week_max))
    
    if (print) {print(summary(lmm))}
    
    return(lmm)
}


add_lmer_row <- function(table, lmm, new_table = FALSE) {
  lmm_tidy <- select(tidy(lmm)[4, ], -c("effect", "group", "term"))
  
  scale <- toupper(gsub("_total", "", out))
  
  # Calculate effect size
  cohensd <- t_to_d(lmm_tidy$statistic, lmm_tidy$df) 
  cohensd_wci <- paste0(round(log(week_max+1)*cohensd$d, 2), 
                        " [", 
                        round(log(week_max+1)*cohensd$CI_low, 2), 
                        ", ", 
                        round(log(week_max+1)*cohensd$CI_high, 2), 
                        "]")

  # Put it together in a row
  if(new_table == TRUE){
    new_table <- cbind("Scale" = scale, 
                       lmm_tidy,
                       "Cohen's d" = cohensd_wci)
  } else {
    new_table <- add_row(table, cbind("Scale" = scale, 
                                      lmm_tidy,
                                      "Cohen's d" = cohensd_wci))   
  }
  return(new_table)
}

table_lmer_to_ft <- function(table_lmer) {
  table_lmer %>%
    mutate('Scale' = gsub("_NEG", " negative", .data[['Scale']]),
           'Scale' = gsub("_POS", " positive", .data[['Scale']])
           ) %>%
    # Give appropriate header names
    rename(
      Estimate = estimate,
      SE = std.error,
      t = statistic,
      DF = df,
      P = p.value) %>%
    mutate(
      across(Estimate:DF, ~ round(., 2)),
      # Format p to three digits and make fat if < .05
      P = round(P, 3)
      ) %>%
    flextable() %>%
    bold(~ P < 0.05, 6) %>%
    autofit()
}

outcome_scales = c("hdrs_total", "madrs_total",
                       "gds_total", "panas_neg_total",
                       "panas_pos_total", "ymrs_total",
                       "CGI1")
```


Results until week 6
```{r results-table-6}
week_max = 6
# Run models and create dataframe with results
for (out in outcome_scales) {
  model <- paste0("lmer.", gsub("_total", "", out))
  
  assign(model, run_lmer(dss, out, week_max))

  # Create table if first predictor...
  if (out == outcome_scales[1]) {
    table_lmer <- add_lmer_row(lmm = get(model), new_table = TRUE)
    # ... else add a row
  } else {
    table_lmer <- add_lmer_row(table = table_lmer, lmm = get(model))
  }
}
      
# Create flextable
ft_lmer <- table_lmer_to_ft(table_lmer) 

save_as_docx(ft_lmer, path = here("Output", "Idosos_MainResults_Week6.docx"))

ft_lmer
```
*Note:* TMA results are based 84 patients and on 2 evaluations only, baseline and week 4. 

Results until week 12
```{r results-table-12}
week_max = 12
# Run models and create dataframe with results
for (out in outcome_scales) {
  model <- paste0("lmer.", gsub("_total", "", out))
  
  assign(model, run_lmer(dss, out, week_max))

  # Create table if first predictor...
  if (out == outcome_scales[1]) {
    table_lmer <- add_lmer_row(lmm = get(model), new_table = TRUE)
    # ... else add a row
  } else {
    table_lmer <- add_lmer_row(table = table_lmer, lmm = get(model))
  }
}
      
# Create flextable
ft_lmer <- table_lmer_to_ft(table_lmer) 

save_as_docx(ft_lmer, path = here("Output", "Idosos_MainResults_Week12.docx"))

ft_lmer
```

## Linear mixed models on per-protocol (PP) sample
```{r}
# Per protocol analyses
dss %>%
  group_by(time) %>%
  summarize(sum(!is.na(hdrs_total)))

subs_pp <- dss %>%
  filter(time < 6) %>%
  group_by(subject) %>%
  mutate(no_missings = sum(is.na(hdrs_total))) %>%
  filter(time == 0, no_missings == 0) %>%
  pull(subject)

dss_pp <- dss %>%
  filter(subject %in% subs_pp)

dss_pp %>%
  group_by(condition) %>%
  filter(time == 0) %>%
  summarize(N = sum(!is.na(hdrs_total)))
```

Results until week 6
```{r results-table-6-pp}
week_max = 6
# Run models and create dataframe with results
for (out in outcome_scales) {
  model <- paste0("lmer.", gsub("_total", "", out))
  
  assign(model, run_lmer(dss_pp, out, week_max))

  # Create table if first predictor...
  if (out == outcome_scales[1]) {
    table_lmer <- add_lmer_row(lmm = get(model), new_table = TRUE)
    # ... else add a row
  } else {
    table_lmer <- add_lmer_row(table = table_lmer, lmm = get(model))
  }
}
      
# Create flextable
ft_lmer <- table_lmer_to_ft(table_lmer) 

save_as_docx(ft_lmer, path = here("Output", "Idosos_MainResults_PP_Week6.docx"))

ft_lmer
```
*Note:* TMA results are based 84 patients and on 2 evaluations only, baseline and week 4. 

Results until week 12
```{r results-table-12-pp}
week_max = 12
# Run models and create dataframe with results
for (out in outcome_scales) {
  model <- paste0("lmer.", gsub("_total", "", out))
  
  assign(model, run_lmer(dss_pp, out, week_max))

  # Create table if first predictor...
  if (out == outcome_scales[1]) {
    table_lmer <- add_lmer_row(lmm = get(model), new_table = TRUE)
    # ... else add a row
  } else {
    table_lmer <- add_lmer_row(table = table_lmer, lmm = get(model))
  }
}
      
# Create flextable
ft_lmer <- table_lmer_to_ft(table_lmer) 

save_as_docx(ft_lmer, path = here("Output", "Idosos_MainResults_PP_Week12.docx"))

ft_lmer
```


```{r, include=FALSE}
## PP, no-log time: no results, but trend: 0.05 < P's < 0.06
model_hdrs_pp_6nolog <- lmer(hdrs_total ~ condition*time + (1|subject), 
                           data = filter(dss_pp, time <= 6))
summary(model_hdrs_pp_6nolog)

model_hdrs_pp_12nolog <- lmer(hdrs_total ~ condition*time + (1|subject), 
                            data = dss_pp)
summary(model_hdrs_pp_12nolog)
```


## LMM with Autoregressive covariance structure (HDRS only)
An autoregressive covariance structure is a type of within-subject correlation structure sometimes used in longitudinal data analysis. It assumes that the correlation between two observations within the same subject decreases as the time between the observations increases.

Until week 6:
```{r autoregressive-covariance-structure-w6}

model_corcar1 <- lme(hdrs_total ~ condition*time, 
                     random = ~ 1 | subject, 
                     #  apply AR1 structure to "time" within each subject, assuming the 
                     correlation = corCAR1(form = ~ time | subject), 
                     data = filter(dss, time <=6),
                    na.action = na.omit)  # na.pass leads to error!

summary(model_corcar1)
```

Until week 12:
```{r autoregressive-covariance-structure-w12}
library(nlme)

model_corcar1 <- lme(hdrs_total ~ condition*time, 
                     random = ~ 1 | subject, 
                     #  apply AR1 structure to "time" within each subject, assuming the 
                     correlation = corCAR1(form = ~ time | subject), 
                     data = dss,
                    na.action = na.omit)  # na.pass leads to error!

summary(model_corcar1)
```

These models show no significant treatment effect ("conditionActive TBS:time").

## Delta analyses for HDRS
These are t-tests that compare the differences between baseline and endpoint for 
the 2 groups. They are a more traditional approach, but only don't take all
measures into account and have less power.  

```{r deltas, message = FALSE}

ttests_table <- tidy(t.test(diff_hdrs_2 ~ condition, dss_wide)) %>%
  rbind(tidy(t.test(diff_hdrs_6 ~ condition, dss_wide))) %>%
  rbind(tidy(t.test(diff_hdrs_12 ~ condition, dss_wide))) %>%
  mutate(
    Endpoint = c("Week 2", "Week 6", "Week 12"),
    across(c(statistic, parameter), ~ round(.x, 2)),
    p.value = round(p.value, 3)
  ) %>% 
  select(Endpoint, t = statistic, DF = parameter, P = p.value) 

ttests_ft <- ttests_table %>%
  flextable() %>%
  # align(j = "p", align = "left", part = "all") %>%
  bold(~ P < 0.05, 4) %>%
  autofit()

ttests_ft
```

## Respone and Remission 
```{r resp-rem-fcts}
resp_rem_outcomes = c("hdrs_response", "hdrs_remission", "madrs_response", "madrs_remission")

# For mixed logistic regression
fit_glmer <- function(df, outcome, week_max, print = FALSE) {
    formula <- as.formula(paste(outcome, "~ condition*time_log + (1|subject)"))

    glmer_model <- glmer(formula, filter(df, time <= week_max), family = binomial)
    
    if (print) {print(summary(glmer_model))}
    
    return(glmer_model)
}


add_glmer_row <- function(table, glmer_model, new_table = FALSE) {

  glmer_tidy <- select(
    tidy(glmer_model, exponentiate = TRUE)[4, ], 
    -c("effect", "group", "term", "std.error"))

  scale <- paste(toupper(word(gsub("_", " ", out), 1)), 
                 word(gsub("_", " ", out), 2))

  # Put it together in a row
  if(new_table == TRUE){
    new_table <- cbind("Scale" = scale,
                       glmer_tidy)
  } else {
    new_table <- add_row(table, cbind("Scale" = scale,
                                      glmer_tidy))
  }
  return(new_table)
}


# For logistic regression
fit_glm <- function(df, outcome, week, print = FALSE) {
    formula <- as.formula(paste(outcome, "~ condition"))

    glm_model <- glm(formula, filter(df, time == week), family = binomial)
    
    if (print) {print(summary(glm_model))}
    
    return(glm_model)
}

add_glm_row <- function(table, glm_model, new_table = FALSE) {

  glm_tidy <- tidy(glm_model, exponentiate = TRUE)[2, ] %>%
    select(-c("term", "std.error"))
  
  scale <- paste(toupper(word(gsub("_", " ", out), 1)), 
                 word(gsub("_", " ", out), 2))

  # Put it together in a row
  if(new_table == TRUE){
    new_table <- cbind("Scale" = scale,
                       glm_tidy)
  } else {
    new_table <- add_row(table, cbind("Scale" = scale,
                                      glm_tidy))
  }
  return(new_table)
}

# For mixed and standard logistic regression
table_glm_to_ft <- function(table_glm) {
  table_glm %>%
    # Give appropriate header names
    rename(
      "Odds ratio" = estimate,
      z = statistic,
      P = p.value) %>%
    mutate(
      across("Odds ratio":z, ~ round(., 2)),
      P = round(P, 3)
      ) %>%
    flextable() %>%
    # align(j = "p", align = "left", part = "all") %>%
    bold(~ P < 0.05, 4) %>%
    autofit()
}

```

### Mixed logistic regressions
This approach considers all time points. Odds ratios indicate the change in 
chances for response or remission for one week of treatment.

Results until week 6
```{r resp-rem-table-6}
week_max = 6

# Run models and create dataframe with results
for (out in resp_rem_outcomes) {
  model <- paste0("glmer.", gsub("_total", "", out))
  
  assign(model, fit_glmer(dss, out, week_max))

  # Create table if first predictor...
  if (out == resp_rem_outcomes[1]) {
    table_glmer <- add_glmer_row(glm = get(model), new_table = TRUE)
    # ... else add a row
  } else {
    table_glmer <- add_glmer_row(table = table_glmer, glm = get(model))
  }
}
      
# Create flextable
ft_glmer <- table_glm_to_ft(table_glmer) 

ft_glmer
```

Results until week 12
```{r resp-rem-table-12}
week_max = 12

# Run models and create dataframe with results
for (out in resp_rem_outcomes) {
  model <- paste0("glmer.", gsub("_total", "", out))
  
  assign(model, fit_glmer(dss, out, week_max))

  # Create table if first predictor...
  if (out == resp_rem_outcomes[1]) {
    table_glmer <- add_glmer_row(glm = get(model), new_table = TRUE)
    # ... else add a row
  } else {
    table_glmer <- add_glmer_row(table = table_glmer, glm = get(model))
  }
}
      
# Create flextable
ft_glmer <- table_glm_to_ft(table_glmer) 

ft_glmer
```

### Logistic regression
Logistic regression only considers data from the endpoint, it's the more common
approach for this type of analysis though. 

Results for week 6
```{r lr-6}
week = 6

# Run models and create dataframe with results
for (out in resp_rem_outcomes) {
  model <- paste0("glm.", gsub("_total", "", out))
  
  assign(model, fit_glm(dss, out, week))

  # Create table if first predictor...
  if (out == resp_rem_outcomes[1]) {
    table_glm <- add_glm_row(glm = get(model), new_table = TRUE)
    # ... else add a row
  } else {
    table_glm <- add_glm_row(table = table_glm, glm = get(model))
  }
}
      
# Create flextable
ft_glm <- table_glm_to_ft(table_glm) 

ft_glm
```

Results for week 12
```{r lr-12}
week = 12

# Run models and create dataframe with results
for (out in resp_rem_outcomes) {
  model <- paste0("glm.", gsub("_total", "", out))
  
  assign(model, fit_glm(dss, out, week))

  # Create table if first predictor...
  if (out == resp_rem_outcomes[1]) {
    table_glm <- add_glm_row(glm = get(model), new_table = TRUE)
    # ... else add a row
  } else {
    table_glm <- add_glm_row(table = table_glm, glm = get(model))
  }
}
      
# Create flextable
ft_glm <- table_glm_to_ft(table_glm) 

ft_glm
```

# Adverse events 
t-Test for group difference in number of adverse events during the entire study"
```{r}

# dss_wide %>%
#   group_by(condition) %>%
#   summarise(mean(adverse_effects_total, na.rm=T),
#             sum(!is.na(adverse_effects_total)),
#             mean(adverse_effects_w12, na.rm=T),
#             sum(!is.na(adverse_effects_w12)))
# 
# dss_wide %>%
#   group_by(condition) %>%
#   summarise(mean(adverse_effects_total, na.rm=T),
#             sd(adverse_effects_total, na.rm=T),
#             sum(!is.na(adverse_effects_total)),)

t_test(dss_wide, adverse_effects_total ~ condition)
```

Logistic regression to test for group difference in number of adverse events for the final week
```{r}
summary(glm(adverse_effects_w12 ~ condition, family = binomial,
    data = dss_wide))
```

# Blinding 
```{r blinding}
# blinding of patients
table_guesses <- as.data.frame.matrix(
  table(dss_wide$condition_rev, dss_wide$guess_patient)
  ) %>%
  rownames_to_column(var = "Guess") %>%
  add_row(Guess = "Patients", .before=1)

table_guesses_alt <- as.data.frame.matrix(
  table(dss_wide$guess_patient, dss_wide$condition_rev)
) %>%
  rownames_to_column(var = "Guess") %>%
  add_row(Guess = "Patients", .before=1)

# chi-square
chisq_patients <- chisq.test(table(dss_wide$condition_rev, dss_wide$guess_patient))

# # Active group only
# chisq.test(table(dss_wide$guess_patient))


# Blinding of raters
table_guesses <- rbind(
  table_guesses,
  as.data.frame.matrix(
    table(dss_wide$condition_rev, dss_wide$guess_rater)) %>%
  rownames_to_column(var = "Guess") %>%
  add_row(Guess = "Raters", .before=1)
)

table_guesses_alt <- rbind(
  table_guesses_alt,
  as.data.frame.matrix(
    table(dss_wide$guess_rater, dss_wide$condition_rev)) %>%
    rownames_to_column(var = "Guess") %>%
    add_row(Guess = "Raters", .before=1)
)



chisq_raters <- chisq.test(table(dss_wide$condition_rev, dss_wide$guess_rater))

# Adding Chi-squares 
table_guesses_chisq <- table_guesses_alt %>%
  mutate(
    `Chi-Square` = case_when(
      row_number() == 3 ~ round(chisq_patients$statistic, 2),
      row_number() == 6 ~ round(chisq_raters$statistic, 2),
      TRUE ~ NA_real_),
    P = case_when(
      row_number() == 3 ~ round(chisq_patients$p.value, 3),
      row_number() == 6 ~ round(chisq_raters$p.value, 3),
      TRUE ~ NA_real_)
  ) %>%
  rename(` ` = Guess)

ft_guesses_chisq <- table_guesses_chisq %>%
  flextable() %>%
  italic(i = c(1, 4)) %>%
  autofit()
ft_guesses_chisq

save_as_docx(ft_guesses_chisq, 
             path = here("Output", "Table_Guesses_ChiSquare.docx"))

```




# Treatment response (HDRS)  prediction 
Predictors were evaluated for HDRS changes (deltas) from baseline to endpoint 
using linear regression and for HDRS response using logistic regression 
(generalized linear model). For each predictor, separate models were set up with
the outcome referring to Week 6 or 12 as possible endpoint. 
Condition and predictor were included as interacting factors.
A significant main effect of the predictor indicates that the predictor had an
effect on HDRS changes/responses on the whole sample, independent of condition.
A significant interaction of the predictor and condition indicates that the 
effectiveness of TBS depended on the predictor.

Variables not included because of no or only very few cases:
- Hyperthyroidism (no case) 
- Stroke ("demarre", only 3 cases)

## Predicting HDRS change
```{r prediction-variables}
preds = c("gender", "Idade", "white_ethnicity", "anos_estudo", 
          "unemployed", "wage", "relationship", 
          "physical_activity", "hipertensao", #"hipotireoidismo", 
          "mini_28", "mini_69", 
          "Benzodiazepines",
          "in_psychotherapy", 
          "failed_AD_trials", "mini_26", 
          "mini_27")
names = c("Gender", "Age", "Ethnicity (white/non-white)", "Years of education",
          "Unemployed (yes/no)", "Wage (low/high)", "Relationship (yes/no)",           
          "Phystical activity (low/high)", "Hypertension", #"Hypothyroidism",
          "Duration of current episode",  "Refractory depression (yes/no)", 
          "Current benzodiazepine use (yes/no)",
          "Currently in psychotherapy (yes/no)", 
          "No. failed AD trials in lifetime", "Onset age of MDD *", 
          "No. previous depressive episodes *")
weeks <- c(6, 12)
```

```{r predict-change}
# Create empty list that will contain models
pred_lms <- list(Week_2 = list(), Week_6 = list(), Week_12 = list(),
                 pred_names = names)

# Create empty df that will contain model stats
pred_lm_table <- tibble(
  Predictor = character(),
  `Main effect::Estimate` = numeric(),
  `Main effect::SE` = numeric(),
  `Main effect::t` = numeric(),
  `Main effect::P` = numeric(),
  `Interaction::Estimate` = numeric(),
  `Interaction::SE` = numeric(),
  `Interaction::t` = numeric(),
  `Interaction::P` = numeric()
)

# Loop over weeks and predictors, fit models and add info to list and df
for (week in weeks) {
  for (i in seq_along(preds)) {
    # Create formula
    current_formula <- as.formula(
      paste0("diff_hdrs_", week, " ~ condition * ", preds[i])
    )
    
    # Fit model
    current_model <- lm(current_formula, dss_wide)
    # Add model to list of models
    pred_lms[[paste0("Week_", week)]][[preds[i]]] <- current_model
    # Create tidy version of model
    current_tidy <- tidy(current_model) %>% select(estimate:p.value)
    
    # Add week indicator if it changes
    if (i == 1) {
      pred_lm_table <- pred_lm_table %>%
        add_row(Predictor = paste("Change to week", week)) 
    }

    # Add current model stats to output dataframe
    pred_lm_table <- rbind(
      pred_lm_table, 
      setNames(
        c("Predictor" = names[i], current_tidy[3,], current_tidy[4,]),
        names(pred_lm_table)
          )
      )
    }
  }

# Transform model output df to flextable
pred_lm_ft <- pred_lm_table %>%
  mutate(
    across(
      c(`Main effect::Estimate`:`Main effect::t`,
        `Interaction::Estimate`:`Interaction::t`),
      ~ round(., 2)),
    across(
      c(`Main effect::P`, `Interaction::P`),
      ~ round(., 3))
    ) %>%
  flextable() %>%
  bold(~ `Main effect::P` < 0.05, 5) %>%
  bold(~ `Interaction::P` < 0.05, 9) %>%
  italic(~startsWith(Predictor, "Change"), 1) %>%
  span_header(sep = "::") %>%
  align(i = 1, align = 'center' ,part = "header")  %>%
  autofit() 

pred_lm_ft

save_as_docx(pred_lm_ft %>% fit_pagewidth(), path = here("Output" , "Predictors_HDRS_change.docx"))

ft_pred_lm_interactiononly <- pred_lm_table %>%
  select(Predictor, `Interaction::Estimate`:`Interaction::P`) %>%
  filter(str_detect(Predictor, 
                    "Change to week 6|Change to week 12|Gender|Age|Duration of current episode|Refractory depression \\(yes/no\\)|Currently in psychotherapy \\(yes/no\\)|Current benzodiazepine use \\(yes/no\\)")
    ) %>%
  mutate(
    across(
      c(`Interaction::Estimate`:`Interaction::t`),
      ~ round(., 2)),
    across(`Interaction::P`,
      ~ round(., 3))
    ) %>%
  rename(Estimate = `Interaction::Estimate`, SE = `Interaction::SE`,
         t = `Interaction::t`, P = `Interaction::P`) %>%
  flextable() %>%
  bold(~ P < 0.05, 5) %>%
  italic(~startsWith(Predictor, "Change"), 1) %>%
  align(i = 1, align = 'center' ,part = "header")  %>%
  autofit() 

ft_pred_lm_interactiononly
save_as_docx(ft_pred_lm_interactiononly %>% fit_pagewidth(), path = here("Output" , "Predictors_HDRS_change_Pub.docx"))
```
*Note:* (*) indicates that data from the HERAH system is missing (33 of 108 patients).

Age at week 6
```{r}
dss %>%
  filter(time == 6) %>%
  mutate(hdrs_improvement = hdrs_total_bl - hdrs_total) %>%
  ggplot(aes(x = Idade, y = hdrs_improvement, 
             color = condition_rev, fill = condition_rev)) +
  geom_point() +
  geom_smooth(method = "lm", alpha = 0.15) +
  labs(x= "Age", y = "HDRS improvement at week 6") +
  theme_nejm()
```

Psychotherapy at week 12
```{r}
dss %>%
  filter(time == 12, !is.na(in_psychotherapy)) %>%
  mutate(psychotherapy = if_else(in_psychotherapy == "FALSE", "No psychotherapy", "In psychotherapy")) %>%
  group_by(condition_rev, psychotherapy) %>%
  summarize(mean = mean(diff_hdrs_12, na.rm = TRUE),
            SEM = sd(diff_hdrs_12, na.rm = TRUE)/sqrt(sum(!is.na(diff_hdrs_12)))
  ) %>%
  ggplot(aes(x = psychotherapy, y = mean, 
             color = condition_rev, fill = condition_rev)) +
  geom_line(aes(group = condition_rev), position = position_dodge(width = 0.1), size = 1) +
  geom_point(position = position_dodge(width = 0.1), size = 3.5) +
  geom_errorbar(position = position_dodge(width = 0.1), 
                aes(ymin = mean - 1.96*SEM,
                    ymax = mean + 1.96*SEM),
                width = 0.7,
                # alpha = 0.8,
                size = 0.9
  ) +
  labs(x= "", y = "HDRS improvement at week 12") +
  theme_nejm()
```


```{r alt-figure}
dss %>%
  filter(time == 12, !is.na(in_psychotherapy)) %>%
  mutate(psychotherapy = if_else(in_psychotherapy == "FALSE", "No psychotherapy", "In psychotherapy")) %>%
  group_by(condition, psychotherapy) %>%
  summarize(mean = mean(diff_hdrs_12, na.rm = TRUE),
            SEM = sd(diff_hdrs_12, na.rm = TRUE)/sqrt(sum(!is.na(diff_hdrs_12)))
  ) %>%
  ggplot(aes(x = condition, y = mean, 
             color = psychotherapy, fill = psychotherapy)) +
  geom_line(aes(group = psychotherapy), position = position_dodge(width = 0.1),size = 1.2) +
  geom_point(position = position_dodge(width = 0.1), size = 3.5) +
  geom_errorbar(position = position_dodge(width = 0.1),
                aes(ymin = mean - 1.96*SEM,
                    ymax = mean + 1.96*SEM),
                width = 0.4,
                size = 0.75,
                alpha = 0.8
  ) +
  labs(x= "Group", y = "HDRS improvement at week 12") +
  scale_color_discrete(direction = -1) +
  theme_nejm()
```

## Predicting HDRS response

```{r predict-response}
# Create empty list that will contain models
pred_glms <- list(Week_2 = list(), Week_6 = list(), Week_12 = list(),
                 pred_names = names)

# Create empty df that will contain model stats
pred_glm_table <- tibble(
  Predictor = character(),
  `Main effect::Odds ratio` = numeric(),
  `Main effect::z` = numeric(),
  `Main effect::P` = numeric(),
  `Interaction::Odds ratio` = numeric(),
  `Interaction::z` = numeric(),
  `Interaction::P` = numeric()
)

# Loop over weeks and predictors, fit models and add info to list and df
for (week in weeks) {
  for (i in seq_along(preds)) {
    # Create formula
    current_formula <- as.formula(
      paste0("hdrs_response_", week, " ~ condition * ", preds[i])
    )
    
    # Fit model
    current_model <- glm(current_formula, binomial, dss_wide)
    # Add model to list of models
    pred_glms[[paste0("Week_", week)]][[preds[i]]] <- current_model

    # Create tidy version of model
    current_tidy <- tidy(current_model, exponentiate = TRUE) %>%
      select(estimate, statistic, p.value)
    
    # Add week indicator if it changes
    if (i == 1) {
      pred_glm_table <- pred_glm_table %>%
        add_row(Predictor = paste("Week", week)) 
    }

    # Add current model stats to output dataframe
    pred_glm_table <- rbind(
      pred_glm_table, 
      setNames(
        c("Predictor" = names[i], current_tidy[3,], current_tidy[4,]),
        names(pred_glm_table)
          )
      )
    }
  }


# Transform model output df to flextable
pred_glm_ft <- pred_glm_table %>%
  mutate(
    across(
      c(`Main effect::Odds ratio`:`Main effect::z`,
        `Interaction::Odds ratio`:`Interaction::z`),
      ~ round(., 2)),
    across(
      c(`Main effect::P`, `Interaction::P`),
      ~ round(., 3))
    ) %>%
  flextable() %>%
  bold(~ `Main effect::P` < 0.05, 4) %>%
  bold(~ `Interaction::P` < 0.05, 7) %>%
  italic(~startsWith(Predictor, "Week"), 1) %>%
  span_header(sep = "::") %>%
  align(i = 1, align = 'center' ,part = "header")  %>%
  autofit()

pred_glm_ft

# save_as_docx(pred_glm_ft %>% fit_pagewidth(), path = "Output/Predictors_HDRS_response.docx")
```
*Note:* (*) indicates that data from the HERAH system is missing (33 of 108 patients).
