==> 3a-secondary-analyses-neuropsychology.Rmd <==
---
title: "TBS for treating depression in elderly: Neuropsychology"
author: "Matthias Lüthi"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    number_sections: true
    toc_float: true
editor_options: 
  markdown: 
    wrap: 72
---


```{r setup, include=FALSE}
# Only show figures and tables:
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
```


```{r load-libraries-data}
library(tidyverse)
library(lme4)
library(lmerTest)
library(lmtest)
library(broom.mixed)
library(effectsize)
library(interactions)
library(flextable)
library(ftExtra)
library(ggpubr)
library(MASS)
library(here)
library(glue)
library(infer)

select <- dplyr::select

# Load custom functions
source("0-custom-functions.R")

load(here("Data", "elderly_neuropsi.RData"))
load(here("Data", "elderly_cleaned_long.RData"))

np <- np_red_lg_ws
remove(np_red_lg_ws)
```


```{r data-cleaning}
np <- np %>%
  mutate(
    subject = as.factor(subject)) %>%
  rename(
  "Intellectual_efficiency" = "Eficiência_intelectual",
  "Attention" = "Atenção",
  "Executive_functions" = "Funções_executivas",
  "Immediate_recall" = "Memória_evocação_imediata",
  "Delayed_recall" = "Memória_evocação_tardia",
  "Visual_praxis" = "Praxia_visual",
  "Verbal_fluency" = "Fluência_verbal",
  age = Idade
) %>% 
  left_join(
    dss %>%
      select(subject, time, education = anos_estudo, hdrs_total, hdrs_total_bl),
    by = c("subject", "time")
  )
  
composite_scores_rep <- np %>%
  select(Attention:Verbal_fluency) %>%
  names()

dss <- dss %>%
  right_join(
    np %>%
      filter(time == 0) %>%
      select(subject, Intellectual_efficiency, all_of(composite_scores_rep)),
    by = "subject"
  )

# Additional DFs
dss_bl <- filter(dss, time == 0)

np_diff <- np %>%
  group_by(subject) %>%
    mutate(
      across(
        c(hdrs_total, all_of(composite_scores_rep)),
        ~ .[time == 4] - .[time == 0],
        .names = "diff4_{.col}"
      )
  ) %>%
  ungroup() %>%
  filter(time == 0) %>%
  left_join(
    dss %>%
      group_by(subject) %>%
      mutate(
        diff6_hdrs_total = hdrs_total[time == 6] - hdrs_total[time == 0],
        diff12_hdrs_total = hdrs_total[time == 12] - hdrs_total[time == 0]
        ) %>%
      ungroup() %>%
      filter(time == 0) %>%
      select(subject, diff6_hdrs_total, diff12_hdrs_total),
    by = "subject")

np4 <- np %>%
  filter(time < 5) 

summaries <- np %>%
  filter(time < 5) %>%
  group_by(time, condition_rev) %>%  
  summarise(across(composite_scores_rep,
                   list(Mean = ~mean(.x, na.rm = TRUE), 
                        SD = ~sd(.x, na.rm = TRUE), 
                        SEM = ~sd(.x, na.rm = TRUE)/sqrt(sum(!is.na(.x))),
                        N = ~sum(!is.na(.x))
                   ),
                   .names = "{col}_{fn}")
  ) %>%
  ungroup()
```

# Figures & Tables
All outliers in neuropsychological test measures have been winsorized: Values 
below the 5th and above the 95th percentile have been replaced by the 5th and 
95th percentile, respectively. 

## Table 1 
```{r table-1, message = FALSE}

# Prepare data frame
table1 <- data.frame(matrix(nrow = 0, ncol = 4))
conditions_with_ns <- mapply(function(x, y) paste0(x, "---N = ", y), 
                             levels(dss_bl$condition_rev), 
                             dss_bl %>%
                               count(condition_rev) %>%
                               pull(n))
colnames(table1) <- c("Characteristic",
                      conditions_with_ns, 
                      "P")

sample_ns <- dss_bl %>% count(condition_rev) %>% pull(n) %>% unique()

vars <- list(c("Demographics", "", "header row"),
             c("gender", "Gender (% Female)", "cat", "female"),
             c("Idade", "Age, y", "cont"),
             c("etnia", "White - no. (%)", "cat", 1),
             c("anos_estudo", "Years of education - mean (SD)", "cont"),
             c("renda", "At least 3 minimum wages - no. (%)", "cat", 2, 3, 4),
             c("ocupacao", "Retired - no. (%)", "cat", 4),
             # c("relationship", "Relationship Status (% Partnered)", "cat", "yes"),
             # c("physical_activity", "Physical Activity (% Active)", "cat", "high"),
             c("Clinical characteristics", "", "header row"),
             c("hipertensao", "Hypertension - no. (%)", "cat", 1),
             c("hipotireoidismo", "Hypothyroidism - no. (%)", "cat", 1),
             c("mini_28", "Duration of current episode, months - mean (SD)", "cont"),
             c("mini_69", "Refractory depression - no. (%)", "cat", "1"),
             # c("in_psychotherapy", "Currently in psychotherapy - no. (%)", "cat", TRUE),
             # c("Benzodiazepines", "Current benzodiazepine use - no. (%)", "cat", "1"),
             # c("failed_AD_trials", "No. failed AD trials in lifetime - median (IQR)", "count"),
             c("mini_26", " Onset age of MDD, years - mean (SD)", "cont"),
             c("mini_27", "No. previous depressive episodes - median (IQR)", "count"),
             c("Scales at baseline", "", "header row"),
             c("hdrs_total", "Hamilton Depression Rating Scale (HDRS-17)", "cont"),
             c("madrs_total", "Montgomery-Asberg Depression Rating Scale (MADRS)", "cont"),
             c("gds_total", "Geriatric Depression Scale (GDS)", "cont"),
             c("ymrs_total", "Young Mania Rating Scale (YMRS)", "cont"),
             c("panas_pos_total", "Positive affect (PANAS)", "cont"),
             c("panas_neg_total", "Negative affect (PANAS)", "cont"),
             c("CGI1", "Clinical Global Impression (CGI)", "cont"),
             c("G_CIRS_0", "Cumulative Illness (CIRS-G)", "cont"),
             c("Composite scores at baseline", "", "header row"),
             c("Intellectual_efficiency", "Intellectual efficiency", "cont"),
             c("Attention", "Attention", "cont"),
             c("Executive_functions", "Executive functions", "cont"),
             c("Immediate_recall", "Immediate recall", "cont"),
             c("Delayed_recall", "Delayed recall", "cont"),
             c("Visual_praxis", "Visual praxis", "cont"),
             c("Verbal_fluency", "Verbal fluency", "cont")
)

for (var in vars) {
  var_name <- var[1]
  var_desc <- var[2]
  var_type <- var[3]

  if (grepl(var_type, "header row")) {
    table1[nrow(table1)+1, ] <- c(var_name, "", "", NA_real_)
    
  } else if (grepl(var_type, "cont")) {

    ns <- dss_bl %>%
      group_by(condition_rev) %>%
      summarize(desc = paste0(round(mean(.data[[var_name]], na.rm=TRUE), 1),
                              " ± ",
                              round(sd(.data[[var_name]], na.rm=TRUE), 1),
                              ifelse(
                                sum(!is.na(.data[[var_name]])) %in% sample_ns,
                                "",
                                paste0("; n=", sum(!is.na(.data[[var_name]]))))
      )) %>%
      pull(desc)

    formula <- as.formula(paste(var_name, "~ condition"))

    table1[nrow(table1)+1, ] <- c(
      var_desc,
      ns,
      tidy(aov(formula, dss_bl))$p.value[1]
    )
    
  } else if (grepl(var_type, "cat")) {
    
    relevant_cats <- var[4:length(var)] 
    el_temp <- dss_bl %>%
      mutate(current_var = if_else(is.na(.data[[var_name]]), NA,
                                  .data[[var_name]] %in% relevant_cats)
             )%>%
      select(condition_rev, current_var)
  
    ns <- el_temp %>%
      group_by(condition_rev) %>%
      summarize(desc = paste0(sum(current_var, na.rm=TRUE),
                              " (",
                              round(sum(current_var, na.rm=TRUE)/sum(!is.na(current_var))*100),
                              ")",
                              ifelse(
                                sum(!is.na(current_var)) %in% sample_ns,
                                "",
                                paste0("; n=", sum(!is.na(current_var))))
      )) %>%
      pull(desc)

    table1[nrow(table1)+1, ] <- c(
      var_desc,
      ns,
      test_independence(el_temp, "condition_rev", "current_var")
    )

  } else if (grepl(var_type, "count")) {
  
    ns <- dss_bl %>%
      group_by(condition_rev) %>%
      summarize(desc = paste0(median(.data[[var_name]], na.rm=TRUE),
                              " (",
                              round(quantile(.data[[var_name]], na.rm=TRUE, probs = 0.25)),
                              "-",
                              round(quantile(.data[[var_name]], na.rm=TRUE, probs = 0.75)),
                              ")",
                              ifelse(
                                sum(!is.na(.data[[var_name]])) %in% sample_ns,
                                "",
                                paste0("; n=", sum(!is.na(.data[[var_name]]))))
      )
      ) %>%
      pull(desc)

    table1[nrow(table1)+1, ] <- c(
      var_desc,
      ns,
      test_count(dss_bl, var_name, "condition_rev")
    )
  }
}

ft_table1  <- table1 %>%
  mutate(P = ifelse(P == "-", 100, round(as.numeric(P), 3)),
         P = case_when(
           P == 100 ~ "-",
           P == 1 ~ ">0.99",
           P == 0 ~ "<0.001",
           P > 0.02 ~ as.character(round(P, 2)),
           TRUE ~ as.character(P))
  ) %>%
  flextable() %>%
  separate_header(sep = "---") %>%
  italic(i = is.na(table1$P), j = 1) %>%
  autofit()
ft_table1

ft_table1_pub  <- table1 %>%
  select(-P) %>%
  flextable() %>%
  italic(i = is.na(table1$P), j = 1) %>%
  separate_header(sep = "---") %>%
  fit_pagewidth()

save_as_docx(
  ft_table1, 
  path = here("Output", "Neuropsychology","Table1_Idosos_NP.docx"))
save_as_docx(
  ft_table1_pub, 
  path = here("Output", "Neuropsychology", "Table1_Idosos_NP_Pub.docx"))

ft_table1
```

## Score table
```{r}
tb_scores <- summaries %>%
  mutate(
    Attention = paste(round(Attention_Mean, 2), "±", round(Attention_SD, 2)),
    Executive_functions = paste(round(Executive_functions_Mean, 2), "±", round(Executive_functions_SD, 2)),
    Immediate_recall = paste(round(Immediate_recall_Mean, 2), "±", round(Immediate_recall_SD, 2)),
    Delayed_recall = paste(round(Delayed_recall_Mean, 2), "±", round(Delayed_recall_SD, 2)),
    Visual_praxis = paste(round(Visual_praxis_Mean, 2), "±", round(Visual_praxis_SD, 2)),
    Verbal_fluency = paste(round(Verbal_fluency_Mean, 2), "±", round(Verbal_fluency_SD, 2))
  ) %>%
  select(time, condition_rev, Attention:Verbal_fluency) %>%
  pivot_longer(cols = Attention:Verbal_fluency,
               names_to = "Scale",
               values_to = "Score"
  ) %>%
  pivot_wider(names_from = time,
               values_from = Score       
  ) %>%
  rename(Baseline = "0", "Week 4" = 4) %>%
  pivot_wider(names_from = condition_rev,
               values_from = c("Baseline", "Week 4"),
               names_sep = "#"         
  )
  

ft_scores <- tb_scores %>%
  flextable() %>%
  span_header(sep = "#") %>%
  align(i = 1, align = 'center', part = "header")  %>%
  autofit() 
ft_scores

save_as_docx(
  ft_scores %>% fit_pagewidth(),
  path = here("Output", "Neuropsychology",
              "Table_CompositeScores.docx"))
```


## Line plots
All error bars represent 95% confidence intervals. 

```{r}
line_plots = list()

for (test in composite_scores_rep) {
  current_mean <- paste0(test, "_Mean")
  current_sem <- paste0(test, "_SEM")
  
  line_plots[[test]] <- summaries %>%
    filter(time != 12) %>%
    ggplot(aes(x = time, y = .data[[current_mean]],
               color = condition_rev, shape = condition_rev)) +
    geom_line(size = 1.5) +
    geom_point(size = 3.5) +
    geom_errorbar(aes(
      ymin = .data[[current_mean]] - 1.96*.data[[current_sem]],
      ymax = .data[[current_mean]] + 1.96*.data[[current_sem]]),
                  width = 0.8,
                  size = 0.75,
                  alpha = 0.8
    ) +
    scale_x_continuous(breaks = c(0, 4),
                       # labels = sapply(c("Base- line", 4),
                       #                 function(x) str_wrap(x, width = 5)),
                       labels = c("Baseline", 4)) +
    expand_limits(y = 0) +
    #scale_color_brewer(palette = "Dark2", direction = 1) +
    labs(y = gsub("_", " ", test), x = "Week") +
    theme_nejm()
  
  # ggsave(here("Output", "Neuropsychology winsorized",  "Composite scores only May 2024",
  #             paste0("Lineplot_ws_", test, ".png")))
  
  print(line_plots[[test]])
}
```
```{r}
# Create combined plot

lp_comb <- ggarrange(
  line_plots[[1]] + labs(x = "") + theme(text = element_text(size = 12)),
  line_plots[[2]] + labs(x = "") + theme(text = element_text(size = 12)),
  line_plots[[3]] + labs(x = "") + theme(text = element_text(size = 12)),
  line_plots[[4]] + labs(x = "") + theme(text = element_text(size = 12)),
  line_plots[[5]] + theme(text = element_text(size = 12)),
  line_plots[[6]] + theme(text = element_text(size = 12)),
  common.legend = TRUE,
  ncol = 2,
  nrow = 3
)
print(lp_comb)

```

# Statistical analysis

## TBS treatment effects on neuropsychological tests
Linear mixed models with covariates ixed effects were time, group, the interaction of time and group, and the covariates age, sex, and education (years of study). 

```{r}

# Create empty list that will contain models
cs_lmms <- list()

# Create empty df that will contain model stats
table_cs_lmm <- tibble(
  Predictor = character(),
  `Effect of time::Estimate` = numeric(),
  `Effect of time::SE` = numeric(),
  `Effect of time::t` = numeric(),
  `Effect of time::DF` = numeric(),
  `Effect of time::P` = numeric(),
  `Treatment effect::Estimate` = numeric(),
  `Treatment effect::SE` = numeric(),
  `Treatment effect::t` = numeric(),
  `Treatment effect::DF` = numeric(),
  `Treatment effect::P` = numeric(),
  .name_repair = "minimal"
)

# Loop over scores, fit models and add info to list and df
for (i in seq_along(composite_scores_rep)) {
  
  # Create formula
  current_formula <- as.formula(
    paste(composite_scores_rep[i], "~ condition*time + age + gender + education + (1|subject)")
  )
  
  # Fit model
  current_model <- lmer(current_formula, np4)
  # Add model to list of models
  cs_lmms[[composite_scores_rep[i]]] <- current_model
  
  
  # Create tidy version of model
  current_tidy <- tidy(current_model) %>% select(estimate:p.value)
  # Add current model stats to output dataframe
  table_cs_lmm <- rbind(
    table_cs_lmm, 
    setNames(
      c("Predictor" = composite_scores_rep[i], current_tidy[3,], current_tidy[7,]),
      names(table_cs_lmm)
    )
  )
}

names(table_cs_lmm) <- gsub("\\.", " ", names(table_cs_lmm))

# Transform model output df to flextable
ft_cs_lmm <- table_cs_lmm %>%
  mutate(
    across(
      c(`Effect of time  Estimate`:`Effect of time  t`,
        `Treatment effect  Estimate`:`Treatment effect  t`),
      ~ round(., 2)),
    across(
      c(`Effect of time  DF`, `Treatment effect  DF`),
      ~ round(., 0)),
    across(
      c(`Effect of time  P`, `Treatment effect  P`),
      ~ ifelse(round(., 3) == 0, "<0.001", round(., 3))),
    Predictor = gsub("_", " ", Predictor)
  ) %>%
  flextable() %>%
  bold(~`Effect of time  P` < 0.05, 6) %>%
  bold(~`Treatment effect  P` < 0.05, 11) %>%
  span_header(sep = "  ") %>%
  align(i = 1, align = 'center', part = "header")  %>%
  align(j = 6, align = "right", part = "all") %>%
  autofit() 

ft_cs_lmm

save_as_docx(
  ft_cs_lmm %>% fit_pagewidth(), 
  path = here("Output", "Neuropsychology", 
  "Table_TreatmentEffects_CompositeScores.docx")
)

```

## Assocations between changes in neuropsychological composite scores and HDRS
```{r}
tb_corrs <- tibble(
  "Composite score" = character(),
  r = numeric(),
  P = numeric()
)

for (score in composite_scores_rep) {
  current_diff <- paste0("diff4_", score)
  current_corr <- cor.test(np_diff$diff4_hdrs_total,
                         np_diff[[current_diff]])
  
  tb_corrs <- tb_corrs %>%
    add_row("Composite score" = score, 
            r = round(current_corr$estimate, 2),
            P = round(current_corr$p.value, 3))
}

ft_corrs <- tb_corrs %>%
  flextable() %>%
  autofit()
ft_corrs
save_as_docx(
  ft_corrs, 
  path = here("Output", "Neuropsychology", 
  "Table_Correlations.docx"))
```


## Neuropsychological tests as predictors of antidepressant TBS response

Prediction of score changes:
These are regression models, with HDRS score change from baseline to week 6 or 
12 as outcome variable, and condition and a neuropsychological test score as 
predictors. The main effect always refers to the neuropsychological test, and 
significance means that this neuropsychological test predicts changes in 
depressive symptoms over time, independent of TBS treatment. A significant 
interaction of test score and condition means that this neuropsychological test 
predicts TBS treatment efficiency. If the interaction is significant, the main
effect cannot be interpreted. 

#### Regressions/linear models
```{r predict-change-excl}
weeks <- c(6, 12)

preds <- c("Intellectual_efficiency", composite_scores_rep) 

# Create empty list that will contain models
pred_lms <- list(Week_6 = list(), Week_12 = list(),
                 pred_names = preds)

# Create empty df that will contain model stats
pred_lm_table <- tibble(
  Predictor = character(),
  Estimate = numeric(),
  SE = numeric(),
  t = numeric(),
  P = numeric(),
  .name_repair = "check_unique"
)

# Loop over weeks and predictors, fit models and add info to list and df
for (week in weeks) {
  current_diff <- paste0("diff", week, "_hdrs_total")
  
  for (i in seq_along(preds)) {

    # Create formula
    current_formula <- as.formula(
      paste0(current_diff, " ~ condition * ", preds[i])
    )
    
    # Fit model
    current_model <- lm(current_formula, np_diff)
    # Add model to list of models
    pred_lms[[paste0("Week_", week)]][[preds[i]]] <- current_model
    # Create tidy version of model
    current_tidy <- tidy(current_model) %>% select(estimate:p.value)

    # Add week indicator if it changes
    if (i == 1) {
      pred_lm_table <- pred_lm_table %>%
        add_row(Predictor = paste("Week", week)) 
    }

    # Add current model stats to output dataframe
    pred_lm_table <- rbind(
      pred_lm_table, 
      setNames(
        c("Predictor" = preds[i], current_tidy[4,]),
        names(pred_lm_table)
          )
      )
    }
  }

# Transform model output df to flextable
pred_lm_ft <- pred_lm_table %>%
  mutate(
    across(
      Estimate:t,
      ~ round(., 2)),
    P = round(P, 3),
    Predictor = gsub("_", " ", Predictor)
    ) %>%
  flextable() %>%
  bold(~P < 0.05, 5) %>%
  italic(~startsWith(Predictor, "Week"), 1) %>%
  autofit() 

pred_lm_ft

save_as_docx(
  pred_lm_ft %>% fit_pagewidth(), 
  path = here("Output", "Neuropsychology", 
              "Predictors_HDRS_change_Reg.docx")
)
```
#### Linear mixed models
```{r}
weeks <- c(6, 12)

preds <- c("Intellectual_efficiency", composite_scores_rep) 

# Create empty list that will contain models
pred_lmms <- list(Week_6 = list(), Week_12 = list(),
                 pred_names = preds)

# Create empty df that will contain model stats
pred_lmm_table <- tibble(
  Predictor = character(),
  Estimate = numeric(),
  SE = numeric(),
  t = numeric(),
  DF = numeric(),
  P = numeric(),
  .name_repair = "check_unique"
)

# Loop over weeks and predictors, fit models and add info to list and df
for (week in weeks) {

  for (i in seq_along(preds)) {
    
    # Create formula
    current_formula <- as.formula(paste0(
      "hdrs_total ~ ", preds[i], " * condition * time + Idade + gender + anos_estudo + (1|subject)")
    )
    
    # Fit model
    current_model <- lmer(current_formula, filter(dss, time <= week))
    # Add model to list of models
    pred_lmms[[paste0("Week_", week)]][[preds[i]]] <- current_model
    # Create tidy version of model
    current_tidy <- tidy(current_model)[11, ] %>% select(estimate:p.value)
    
    # Add week indicator if it changes
    if (i == 1) {
      pred_lmm_table <- pred_lmm_table %>%
        add_row(Predictor = paste("Week", week)) 
    }
    
    # Add current model stats to output dataframe
    pred_lmm_table <- rbind(
      pred_lmm_table, 
      setNames(
        c("Predictor" = preds[i], current_tidy),
        names(pred_lmm_table)
      )
    )
  }
}

# Transform model output df to flextable
pred_lmm_ft <- pred_lmm_table %>%
  mutate(
    across(
      Estimate:t,
      ~ round(., 2)),
    P = ifelse(round(P, 3) == 0, "<0.001", round(P, 3)),
    DF = round(DF, 0),
    Predictor = gsub("_", " ", Predictor)
  ) %>%
  flextable() %>%
  bold(~P < 0.05, 6) %>%
  italic(~startsWith(Predictor, "Week"), 1) %>%
  align(j = 6, align = "right", part = "all") %>%
  autofit() 

pred_lmm_ft

save_as_docx(
  pred_lmm_ft %>% fit_pagewidth(), 
  path = here("Output", "Neuropsychology", "Predictors_HDRS_change_LMM.docx")
)

```

#### Figures for significant predictors at week 12

```{r fig.height=8, fig.width=9}
g1 <- ggplot(np_diff, aes(Intellectual_efficiency, -diff12_hdrs_total, 
                 color = condition, fill = condition)) + 
  geom_point() + 
  geom_smooth(method = "lm") + 
  theme_nejm() +
  labs(x = "Intellectual efficiency", y = "HDRS-17 improvement\nuntil week 12")

g2 <- ggplot(np_diff, aes(Attention, -diff12_hdrs_total, 
                 color = condition, fill = condition)) + 
  geom_point() + 
  geom_smooth(method = "lm") + 
  theme_nejm() +
  labs(x = "Attention", y = "")

g3 <- ggplot(np_diff, aes(Executive_functions, -diff12_hdrs_total, 
                 color = condition, fill = condition)) + 
  geom_point() + 
  geom_smooth(method = "lm") + 
  theme_nejm() +
  labs(x = "Executive functions", y = "")
g4 <- ggplot(np_diff, aes(Immediate_recall, -diff12_hdrs_total, 
                 color = condition, fill = condition)) + 
  geom_point() + 
  geom_smooth(method = "lm") + 
  theme_nejm() +
  labs(x = "Immediate recall", y = "HDRS-17 improvement\nuntil week 12")

g5 <- ggplot(np_diff, aes(Delayed_recall, -diff12_hdrs_total, 
                 color = condition, fill = condition)) + 
  geom_point() + 
  geom_smooth(method = "lm") + 
  theme_nejm() +
  labs(x = "Delayed recall", y = "")

ggarrange(g1, g2, g3, g4, g5, common.legend = T)

ggsave(here("Output", "Neuropsychology", "SignificantPredictors.png"))


```
